modify app.py and rag.py
